Add mesh colors and change to using floats for most things

+++ src/main/java/io/github/kale_ko/gighm/rendering/Window.java
@@ -294,7 +294,7 @@ public class Window {
             renderer.render();
 
             Instant now = Instant.now();
-            Double delta = (now.getEpochSecond() + ((double) now.getNano() / 1000000000)) - (lastRender.getEpochSecond() + ((double) lastRender.getNano() / 1000000000));
+            Float delta = (float) ((now.getEpochSecond() + ((double) now.getNano() / 1000000000)) - (lastRender.getEpochSecond() + ((double) lastRender.getNano() / 1000000000)));
 
             this.eventManager.emit(new RenderEvent(delta));
